Factor out common MachineNot*Exception logic
